adicionando novo problema ao dojo de cucumber.

+++ cucumber/servers_and_queue_consumer/servers/server.rb
@@ -2,7 +2,11 @@ require "sinatra"
 
 get '/api/:name' do
   content_type 'plain/text'
-  content = File.open("files/#{params[:name]}").read
-  puts content
-  content
+
+  puts File.exists?("files/#{params[:name]}")
+  if File.exists?("files/#{params[:name]}")
+    File.open("files/#{params[:name]}").read
+  else
+    status 404
+  end
 end
